Make seperate edit components
