+++ client/src/component/Pages/students/Addstd.jsx
@@ -1,6 +1,7 @@
-import { useState, useRef } from 'react';
+import { useState } from 'react';
 import { useNavigate } from 'react-router';
 import React, { useEffect } from 'react';
+import Edit from './Editform';
 
 const Addstd = () => {
   const [activity, setActivity] = useState({
@@ -32,10 +33,6 @@ const Addstd = () => {
   const [update, setUpdate] = useState(-1);
   const [error, showError] = useState('');
 
-  const nameRef = useRef();
-  const ageRef = useRef();
-  const sexRef = useRef();
-  const phoneRef = useRef();
 
   // event handler for input change
   function handleChange(e) {
@@ -114,77 +111,9 @@ const Addstd = () => {
     setUpdate(i);
   }
 
-  // render the component
-  function Edit({ current, listData, setListData }) {
-    const [editedData, setEditedData] = useState({ ...current });
-
-    const handleChange = (e) => {
-      const { name, value } = e.target;
-      setEditedData((prevData) => ({ ...prevData, [name]: value }));
-    };
-
-    const handleUpdate = () => {
-      const newList = listData.map((li, index) =>
-        index === current.index ? editedData : li
-      );
-      setListData(newList);
-      setUpdate(-1); // Exit the edit mode
-    };
-
-    return (
-      <tr>
-        <td>
-          <input
-            type="text"
-            name="name"
-            value={editedData.name}
-            onChange={handleChange}
-            ref={nameRef}
-          />
-        </td>
-        <td>
-          <input
-            type="text"
-            name="age"
-            value={editedData.age}
-            onChange={handleChange}
-            ref={ageRef}
-          />
-        </td>
-        <td>
-          <select
-            placeholder="Choose Gender"
-            onChange={handleChange}
-            value={editedData.sex}
-            ref={sexRef}
-            name="sex"
-          >
-            <option value="">Select Gender</option>
-            <option value="male">Male</option>
-            <option value="female">Female</option>
-          </select>
-        </td>
-        <td>
-          <input
-            type="text"
-            name="phone"
-            value={editedData.phone}
-            onChange={handleChange}
-            ref={phoneRef}
-          />
-        </td>
-        <td>
-          <button type="button" onClick={handleUpdate}>
-            Update
-          </button>
-        </td>
-      </tr>
-    );
-  }
-  
-  function navigateHome() {
-    navigate('/')
-  }
+function navigateHome() {
+  navigate('/')
+}
 
   // const [select, setSelect]  = useState();
   return (
@@ -249,6 +178,9 @@ const Addstd = () => {
             {listData.map((data, index) =>
               update === index ? (
                 <Edit
+                key={index}
+            
+                setUpdate={setUpdate}
                   current={{ ...data, index }}
                   listData={listData}
                   setListData={setListData}
